+++ docker/mysql/data/binlog.index
@@ -1,8 +1,3 @@
-./binlog.000001
-./binlog.000002
-./binlog.000003
-./binlog.000004
-./binlog.000005
 ./binlog.000006
 ./binlog.000007
 ./binlog.000008
@@ -19,3 +14,4 @@
 ./binlog.000019
 ./binlog.000020
 ./binlog.000021
+./binlog.000022
